api-ref/requirements.txt
requests==2.21.0
aiohttp_cors==0.7.0
py_solc==3.2.0
web3==5.6.0
pymongo==3.9.0
aiohttp==3.6.2
PyJWT==1.7.1
PyYAML==5.3.1
solc==0.0.4
